+++ SS2-Project/Assets/Starstorm2/Modules/EntityStates/Drones/CloneDroneCook.cs
@@ -2,7 +2,7 @@
 using RoR2;
 using SS2;
 using UnityEngine.Networking;
-
+using SS2.Interactables;
 namespace EntityStates.CloneDrone
 {
     public class CloneDroneCook : BaseSkillState
@@ -207,7 +207,7 @@ namespace EntityStates.CloneDrone
                 outer.SetNextStateToMain();
                 return;
             }
-            if (duration <= fixedAge)
+            if (fixedAge >= duration)
             {
                 outer.SetNextStateToMain();
                 //create the item
@@ -229,7 +229,16 @@ namespace EntityStates.CloneDrone
 
         public void CreateClone()
         {
-            PickupDropletController.CreatePickupDroplet(gpc.pickupIndex, transform.position, Vector3.up * dropUpVelocityStrength + transform.forward * dropForwardVelocityStrength);
+            GenericPickupController.CreatePickupInfo info = new GenericPickupController.CreatePickupInfo
+            {
+                prefabOverride = CloneDroneDamaged.clonedPickupPrefab,
+                pickupIndex = gpc.pickupIndex,
+                position = transform.position,
+            };
+            var component = this.gpc.gameObject.GetComponent<CloneDroneDamaged.ClonedPickup>();
+            if (!component) component = gpc.gameObject.AddComponent<CloneDroneDamaged.ClonedPickup>();
+            component.OnCloned(base.gameObject);
+            PickupDropletController.CreatePickupDroplet(info, transform.position, Vector3.up * dropUpVelocityStrength + transform.forward * dropForwardVelocityStrength);
         }