+++ SS2-Project/Assets/Starstorm2/Modules/Pickups/Items/Boss/Remuneration.cs
@@ -18,7 +18,7 @@ namespace SS2.Items
             On.RoR2.PickupDisplay.RebuildModel += EnableVoidParticles;
             On.RoR2.PickupDisplay.RebuildModel += EnableVoidParticles;
 
-            deleteEffectPrefab = PrefabAPI.InstantiateClone(Addressables.LoadAssetAsync<GameObject>("RoR2/Base/Nullifier/NullifierExplosion.prefab").WaitForCompletion(), "WAWA");
+            deleteEffectPrefab = PrefabAPI.InstantiateClone(Addressables.LoadAssetAsync<GameObject>("RoR2/Base/Nullifier/NullifierExplosion.prefab").WaitForCompletion(), "WAWA", false);
             deleteEffectPrefab.GetComponent<EffectComponent>().soundName = "Play_nullifier_death_vortex_explode"; // cringe.
         }