+++ SS2-Project/Assets/Starstorm2/ScriptableObjects/NemesisInventory.cs
@@ -30,13 +30,10 @@ namespace SS2
         private static void Initialize()
         {
             AddressReferencedAsset.OnAddressReferencedAssetsLoaded += InitializeNemesisInventories;
-            SS2Log.Info("A NEMESISINVENTORY");
             void InitializeNemesisInventories()
             {
-                SS2Log.Info("B");
                 foreach (NemesisInventory nemesisInventory in instances)
                 {
-                    SS2Log.Info("C");
                     nemesisInventory.CreateWeightedSelection();
                 }
             }